osr/requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-import-export==1.2.0
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
jdcal==1.4.1
lml==0.0.9
MarkupPy==1.14
odfpy==1.4.0
openpyxl==2.5.14
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
